move enums KeylockMode & KeyunlockMode to separate defs file
