+++ src/engine/controls/keycontrol.cpp
@@ -7,15 +7,13 @@
 #include "control/controlpotmeter.h"
 #include "control/controlproxy.h"
 #include "control/controlpushbutton.h"
+#include "engine/defs_keylock.h"
 #include "engine/enginebuffer.h"
 #include "mixer/playermanager.h"
 #include "moc_keycontrol.cpp"
 #include "track/keyutils.h"
 
 constexpr bool kEnableDebugOutput = false;
-
-static const double kLockCurrentKey = 1;
-static const double kKeepUnlockedKey = 1;
 
 KeyControl::KeyControl(const QString& group,
         UserSettingsPointer pConfig)
@@ -147,9 +145,9 @@ void KeyControl::slotRateChanged() {
     updateRate();
 }
 
-// This is called when rate_ratio, vinylcontrol_rate, vinylcontrol_enabled or
-// keylock are changed, but also when EngineBuffer::processTrackLocked requests
-// m_pitchRateInfo struct while rate, pitch or pitch_adjust were just updated.
+/// This is called when rate_ratio, vinylcontrol_rate, vinylcontrol_enabled or
+/// keylock are changed, but also when EngineBuffer::processTrackLocked requests
+/// m_pitchRateInfo struct while rate, pitch or pitch_adjust were just updated.
 void KeyControl::updateRate() {
     if (m_pVCEnabled && m_pVCRate && m_pVCEnabled->toBool()) {
         m_pitchRateInfo.tempoRatio = m_pVCRate->get();
@@ -191,7 +189,9 @@ void KeyControl::updateRate() {
 
     if (m_pKeylock->toBool()) {
         if (!m_pitchRateInfo.keylock) {                    // Enabling keylock
-            if (m_keylockMode->get() == kLockCurrentKey) { // Lock at current pitch
+            if (m_keylockMode->get() ==
+                    static_cast<double>(KeylockMode::LockCurrentKey)) {
+                // Lock at current pitch
                 speedSliderPitchRatio = m_pitchRateInfo.tempoRatio;
                 if constexpr (kEnableDebugOutput) {
                     qDebug() << "   LOCKING current key";
@@ -223,7 +223,7 @@ void KeyControl::updateRate() {
         }
     } else {                           // !m_pKeylock
         if (m_pitchRateInfo.keylock) { // Disabling Keylock
-            if (m_keyunlockMode->get() == kKeepUnlockedKey) {
+            if (m_keyunlockMode->get() == static_cast<double>(KeyunlockMode::KeepLockedKey)) {
                 // adopt speedSliderPitchRatio change as pitchTweakRatio
                 m_pitchRateInfo.pitchTweakRatio *=
                         (speedSliderPitchRatio / m_pitchRateInfo.tempoRatio);
@@ -299,6 +299,8 @@ void KeyControl::updateRate() {
     updateKeyCOs(dFileKey, pitchOctaves);
 }
 
+/// This is called when the file key is changed (by analysis or user input),
+/// and when a track with a different key is loaded
 void KeyControl::slotFileKeyChanged(double value) {
     updateKeyCOs(value,  m_pPitch->get() / 12);
 }
@@ -352,6 +354,9 @@ void KeyControl::setEngineKey(double key, double key_distance) {
     return;
 }
 
+/// This is called when pitch is changed either by user interaction or
+/// when BaseTrackPlayerImpl resets the pitch on track load per configuration
+/// when key is locked with KeylockMode::LockCurrentKey
 void KeyControl::slotPitchChanged(double pitch) {
     Q_UNUSED(pitch)
     m_updatePitchRequest = 1;
@@ -391,6 +396,8 @@ void KeyControl::updatePitch() {
     }
 }
 
+/// This is called when pitch_adjust is changed either by user interaction or
+/// when BaseTrackPlayerImpl resets the pitch on track load per configuration
 void KeyControl::slotPitchAdjustChanged(double pitchAdjust) {
     Q_UNUSED(pitchAdjust);
     m_updatePitchAdjustRequest = 1;
