+++ src/engine/defs_keylock.h
@@ -1,0 +1,11 @@
+#pragma once
+
+enum class KeylockMode {
+    LockOriginalKey,
+    LockCurrentKey
+};
+
+enum class KeyunlockMode {
+    ResetLockedKey,
+    KeepLockedKey
+};
