+++ src/preferences/dialog/dlgprefdeck.h
@@ -4,6 +4,7 @@
 
 #include "engine/controls/cuecontrol.h"
 #include "engine/controls/ratecontrol.h"
+#include "engine/defs_keylock.h"
 #include "preferences/dialog/dlgpreferencepage.h"
 #include "preferences/dialog/ui_dlgprefdeckdlg.h"
 #include "preferences/usersettings.h"
@@ -33,16 +34,6 @@ namespace TrackTime {
         HECTO_SECONDS,
     };
 }
-
-enum class KeylockMode {
-    LockOriginalKey,
-    LockCurrentKey
-};
-
-enum class KeyunlockMode {
-    ResetLockedKey,
-    KeepLockedKey
-};
 
 enum class LoadWhenDeckPlaying {
     Reject,
